backend/requirements.txt
fastapi
fastapi[all]
uvicorn

python-multipart
sqlalchemy
psycopg2-binary
python-dotenv

requests 
google-genai
pillow